% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{achilles}
\alias{achilles}
\title{Runs ACHILLES on AmsterdamUMCdb converted to the OMOP CDM}
\usage{
achilles()
}
\description{
This function runs the Automated Characterization of Health
Information at Large-Scale Longitudinal Evidence Systems (ACHILLES)
Achilles provides descriptive statistics on an OMOP CDM database.
}
