% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_system_settings.R
\name{alter_system_settings}
\alias{alter_system_settings}
\title{Alters PostgreSQL System settings.}
\usage{
alter_system_settings(defaults = FALSE)
}
\arguments{
\item{defaults}{Determines to apply default or optimized PostgreSQL settings. Default: FALSE}
}
\description{
This function alters the PostgreSQL settings to reduce house
keeping while bulk inserting records.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}.

These settings require a server restart.
}
