% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_corrections.R
\name{apply_corrections}
\alias{apply_corrections}
\title{Apply corrections}
\usage{
apply_corrections()
}
\description{
This function applies corrections to the clinical data
based on issues identified using DataQualityDashboard.
}
\details{
This function executes the \href{https://github.com/AmsterdamUMC/AMSTEL/blob/main/inst/sql/apply_corrections.sql}{\code{apply_corrections.sql}}
file. For a list of applied corrections, see the comments in this file.

In order for this function to work, the CDM instance already needs to be
contain clinical data (e.g. after \code{amstel::load_data_tables}).
}
