% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue_export.R
\name{catalogue_export}
\alias{catalogue_export}
\title{Run EHDEN CatalogueExport}
\usage{
catalogue_export()
}
\description{
Runs CatalogueExport to export the data from the OMOP-CDM for inclusion in
the EHDEN Database Catalogue. CatalogueExport executes a
set of analyses based on OHDSI \href{https://github.com/OHDSI/Achilles}{Achilles}.
}
