% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm_inspection.R
\name{cdm_inspection}
\alias{cdm_inspection}
\title{Runs the CdmInspection to check the OMOP-CDM instance before submission
to EHDEN}
\usage{
cdm_inspection()
}
\description{
Runs the CdmInspection to check the OMOP-CDM instance before submission
to EHDEN
}
