% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{cloud_auth}
\alias{cloud_auth}
\title{Small wrapper for authentication https://github.com/ropensci/googleLanguageR.
Allows authentication using Google Cloud API credentials
in default user folder'}
\usage{
cloud_auth()
}
\description{
Small wrapper for authentication https://github.com/ropensci/googleLanguageR.
Allows authentication using Google Cloud API credentials
in default user folder'
}
\examples{
cloud_auth()
}
