% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to database}
\usage{
connect(database = "cdm", default_schema = "")
}
\arguments{
\item{database}{Database to use, either "amsterdamumcdb",  "cdm", "results",
or "temp".
Default: "cdm"}

\item{default_schema}{Default schema to use as the search path. By default,
this will be set to the value of \code{schema} in the configuration section
of \code{database}}
}
\value{
DatabaseConnector Connection object
}
\description{
Connects to the AmsterdamUMCdb or CDM database based on \code{config.yaml} in
the user config folder.
}
\examples{
conn <- connect()
conn <- connect("amsterdamumcdb")
}
