% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cdm_indices.R
\name{create_cdm_indices}
\alias{create_cdm_indices}
\title{Create Indices on Common Data Model Tables.}
\usage{
create_cdm_indices()
}
\description{
This function creates indices for the tables in the CDM.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create the indices for the tables of the Common Data Model.

This function assumes \code{amstel::create_cdm_tables()} has been run successfully.
}
