% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cdm_tables.R
\name{create_cdm_tables}
\alias{create_cdm_tables}
\title{Create Common Data Model tables}
\usage{
create_cdm_tables()
}
\description{
This function creates the required CDM tables using the \code{executeDdl} function
from the \code{OHDSI/CommonDataModel} package.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create the Common Data Model tables using the
\code{executeDdl} function from the \code{OHDSI/CommonDataModel} package.
}
\examples{
create_cdm_tables()
}
