% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort_tables.R
\name{create_cohort_tables}
\alias{create_cohort_tables}
\title{Create the Common Data Model Results Cohort Tables}
\usage{
create_cohort_tables()
}
\description{
This function creates the Common Data Model Cohort tables, than can be used
to define cohorts based on criteria using standard concepts. An example cohort
has been created to select patients receiving mechanical ventilation
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create WebAPI tables.
}
\examples{
create_cohort_tables()
}
