% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_tables.R
\name{create_data_tables}
\alias{create_data_tables}
\title{Create the Common Data Model Clinical Data tables}
\usage{
create_data_tables()
}
\description{
This function creates the Common Data Model Clinical Data tables, e.g.
all tables with the exception of the tables related to the Standard
Vocabularies and Standard Concepts.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create WebAPI tables.
}
\examples{
create_webapi_tables()
}
