% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{create_environment}
\alias{create_environment}
\title{Creates a new database environment (database schemes) for AmsterdamUMCdb
and the CDM data and results schemes for testing and/or
example purposes.}
\usage{
create_environment(name = "test", clean = FALSE)
}
\arguments{
\item{name}{Name of the environment to create.
Default: "test"}

\item{clean}{Specifies whether to start with completely empty schemes by dropping the
existing ones. WARNING: this will completely remove both schemes if the
specified user is the owner of the schemes.
Default: FALSE}
}
\value{
list of connections to both schemes
}
\description{
Creates a new database environment (database schemes) for AmsterdamUMCdb
and the CDM data and results schemes for testing and/or
example purposes.
}
