% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vocabulary_tables.R
\name{create_vocabulary_tables}
\alias{create_vocabulary_tables}
\title{Create the Common Data Model Vocabulary tables}
\usage{
create_vocabulary_tables()
}
\description{
This function creates the Common Data Model Standard Vocabularies and
Standard Concepts tables.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create WebAPI tables.
}
\examples{
create_vocabulary_tables()
}
