% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_webapi_tables.R
\name{create_webapi_tables}
\alias{create_webapi_tables}
\title{Create OHDSI WebAPI tables}
\usage{
create_webapi_tables()
}
\description{
This function creates the required OHDSI WebAPI tables.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create WebAPI tables.
}
\examples{
create_webapi_tables()
}
