% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_drivers.R
\name{download_drivers}
\alias{download_drivers}
\title{Downloads JDBC drivers}
\usage{
download_drivers()
}
\description{
Downloads the required JDBC drivers for connecting to the database
management system(s) as specified by the \code{config.yaml} file.
}
\examples{
amstel::download_drivers()
}
