% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_data_tables.R
\name{drop_data_tables}
\alias{drop_data_tables}
\title{Drop CDM Data Tables.}
\usage{
drop_data_tables(cascade = FALSE)
}
\arguments{
\item{cascade}{Automatically drop objects that depend on the table including other tables,
and in turn all objects that depend on those objects. Default: FALSE.}
}
\description{
This function drops CDM tables, excluding the vocabulary tables.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to delete the Clinical Data tables from the
Common Data Model.
}
