% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_vocabulary_tables.R
\name{drop_vocabulary_tables}
\alias{drop_vocabulary_tables}
\title{Drop Vocabulary Tables.}
\usage{
drop_vocabulary_tables(cascade = FALSE)
}
\arguments{
\item{cascade}{Determines whether to drop dependent objects too.
\strong{WARNING}: Since all CDM data tables depend on the vocabulary,
this will delete most data tables as well. Default: FALSE}
}
\description{
This function drops Vocabulary tables in a CDM.
}
