% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Gets the configuration for working with the amstel package. Reads the
config.yaml file from the user config folder or creates a new file if the
configuration file is missing.}
\usage{
get_config()
}
\value{
list containing configuration settings
}
\description{
Gets the configuration for working with the amstel package. Reads the
config.yaml file from the user config folder or creates a new file if the
configuration file is missing.
}
\examples{
config <- get_config()
}
