% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_server_tablename}
\alias{get_server_tablename}
\title{Get case-sensitive table name on server}
\usage{
get_server_tablename(tablename, conn)
}
\arguments{
\item{tablename}{Name of the table to check}

\item{conn}{DatabaseConnector Connection object}
}
\value{
String server table name
}
\description{
Since some DBMS are case-sensitive use the names of the tables currently
stored in the database to prevent creating new ones with a different case
}
\examples{
\dontshow{if (has_cdm_environment()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
conn <- connect("cdm")
table <- get_server_tablename("PROCEDURE_OCCURRENCE", conn)
table
\dontshow{\}) # examplesIf}
}
