% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{has_adb_environment}
\alias{has_adb_environment}
\title{Checks whether the AmsterdamUMCdb environment is available for running
example code}
\usage{
has_adb_environment()
}
\value{
Bool
}
\description{
Checks whether the AmsterdamUMCdb environment is available for running
example code
}
