% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{has_cdm_environment}
\alias{has_cdm_environment}
\title{Checks whether the CDM environment is available for running example code}
\usage{
has_cdm_environment()
}
\value{
Bool
}
\description{
Checks whether the CDM environment is available for running example code
}
