% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{has_example_environment}
\alias{has_example_environment}
\title{Checks whether an example environment can be used}
\usage{
has_example_environment()
}
\value{
Bool
}
\description{
Checks whether an example environment can be used
}
