% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_vocabulary_tables.R
\name{index_vocabulary_tables}
\alias{index_vocabulary_tables}
\title{Create Indices on Common Data Model vocabulary tables.}
\usage{
index_vocabulary_tables()
}
\description{
This function creates primary keys and indices for the
vocabulary tables in the CDM.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create the indices for the tables of the
Common Data Model. Since the \code{concept} will be used often during the ETL of
the AmsterdamUMCdb tables, it helps when these (static) tables have already
been indexed.

This function assumes \code{amstel::create_cdm_tables()} and
\code{amstel::load_vocabulary_tables()} has been run successfully.
}
