% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_concepts.R
\name{load_concepts}
\alias{load_concepts}
\title{Load concepts}
\usage{
load_concepts(concept_group)
}
\arguments{
\item{concept_group}{concept group name that will be used to read the file}
}
\value{
tibble containing the concepts belonging to this group
}
\description{
Reads a (previously generated) concept list by based on the concept group name
}
\examples{
providers <- load_concepts("providers")
}
