% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_tables.R
\name{load_data_tables}
\alias{load_data_tables}
\title{Load CDM Standardized Clinical Data Tables.}
\usage{
load_data_tables()
}
\description{
This function loads the CDM Clinical Data tables with data from
AmsterdamUMCdb
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{amsterdamumcdb} and \code{cdm} sections to extract, transform and load the
data from AmsterdamUMCdb into the Common Data Model.

This function assumes \code{amstel::create_cdm_tables()},
\code{amstel::load_vocabulary_tables()} and \code{amstel::load_source_to_concept_map()}
have been run successfully.
}
