% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_source_to_concept_map.R
\name{load_source_to_concept_map}
\alias{load_source_to_concept_map}
\title{Load source_to_concept_map table from Usagi mapping Files.}
\usage{
load_source_to_concept_map()
}
\description{
This function loads the source_to_concept_map table using the mapping
files created with the OHDSI Usagi (https://github.com/OHDSI/Usagi) application.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to load the source_to_concept tables in the CDM by reading the
Usagi csv files in the \code{mappings} folder.

This function assumes \code{amstel::create_cdm_tables()} and
\code{amstel::load_vocabulary_tables()} have already been run.
}
\examples{
stcm <- load_source_to_concept_map()
stcm

}
