% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_source_to_value_map.R
\name{load_source_to_value_map}
\alias{load_source_to_value_map}
\title{Load source_to_value_map table from SQL file.}
\usage{
load_source_to_value_map()
}
\description{
Since Usagi removes mappings to invalid concepts on loading
(e.g. after importing a new vocabulary with deprecated concepts),
our previous method to store quantities for Drug and other domains will
fail since many numeric concepts have been deprecated.
This function will load migrated quantity data from the
\code{insert_source_to_value_map.sql} file to the new source_to_value table.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to load the source_to_concept tables in the CDM by reading the
Usagi csv files in the \code{mappings} folder.

This function assumes \code{amstel::create_cdm_tables()} and
\code{amstel::load_vocabulary_tables()} have already been run.
}
\examples{
load_source_to_value_map()

}
