% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_concepts.R
\name{load_usagi_concepts}
\alias{load_usagi_concepts}
\title{Load Usagi Concepts}
\usage{
load_usagi_concepts(concept_group)
}
\arguments{
\item{concept_group}{concept group name that will be used to read the file}
}
\value{
tibble containing the concepts belonging to this group
}
\description{
Reads a (partially mapped) concept list based on the concept group name. This
function assumes the mapped concept list had been stored in the \code{data/mappings}
folder.
}
\examples{
providers <- load_usagi_concepts("providers")
}
