% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vocabulary_tables.R
\name{load_vocabulary_tables}
\alias{load_vocabulary_tables}
\title{Load Vocabulary Tables From CSV Files.}
\usage{
load_vocabulary_tables()
}
\description{
This function loads the vocabulary tables with csv files downloaded
from OHDSI Athena (https://athena.ohdsi.org/vocabulary/list).
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{cdm} section to create the Vocabulary tables in the CDM by reading the
vocabulary csv files downloaded from the Athena from the path as
specified in the \code{data} section.

This function assumes \code{amstel::create_cdm_tables()} has already been run.
}
