% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_source_to_value_map.R
\name{migrate_source_to_value_map}
\alias{migrate_source_to_value_map}
\title{Migrates the source_to_concept_map data to the
new source_to_value_map table}
\usage{
migrate_source_to_value_map()
}
\description{
Migrates the source_to_concept_map data to the
new source_to_value_map table
}
