% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vocabulary.R
\name{read_vocabulary}
\alias{read_vocabulary}
\title{Read vocabulary}
\usage{
read_vocabulary(vocabulary)
}
\arguments{
\item{vocabulary}{Vocabulary name, either the name of the vocabulary table
or name of the file. Case-insensitive.}
}
\value{
Tibble containing data from one of the vocabulary files
}
\description{
Reads the vocabulary from the one of the csv files in the Vocabulary path (
Default: \code{data/vocabulary}). If the file does not exist, it will show the
available csv files in that folder.
}
\examples{
concept <- read_vocabulary("concept")
}
