% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_concepts.R
\name{save_concepts}
\alias{save_concepts}
\title{Save Concepts}
\usage{
save_concepts(concepts, concept_group, force = FALSE)
}
\arguments{
\item{concepts}{tibble containing the concepts}

\item{concept_group}{concept group name that will be used to create the file name}

\item{force}{forces overwriting file. WARNING: this may delete already mapped concepts}
}
\value{
(unmodified) tibble containing the concepts
}
\description{
Saves the concepts list into the source_concepts folder
}
\examples{
\dontshow{if (has_adb_environment()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(dplyr)
conn <- amstel::connect("amsterdamumcdb")
admissions <- tbl(conn, "admissions")
concepts <- admissions \%>\%
  group_by(source_concept=origin) \%>\%
  summarise(count=n()) \%>\%
  collect()
concepts <- save_concepts(concepts, "admissions_origin")
concepts
\dontshow{\}) # examplesIf}
}
