% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_dashboard.R
\name{show_dqd_log}
\alias{show_dqd_log}
\title{Show Data Quality Dashboard log (Shiny App)}
\usage{
show_dqd_log()
}
\description{
Shows the log created while executing
DataQualityDashboard::executeDqChecks()
}
