% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_concepts.R
\name{swap_usagi_counts}
\alias{swap_usagi_counts}
\title{Swap counts}
\usage{
swap_usagi_counts(concept_group)
}
\arguments{
\item{concept_group}{concept group name that will be used to read the file}
}
\value{
tibble containing the usagi mappings with counts swapped
}
\description{
Swaps the sourceFrequency and counts/counts_validated columns to allow
correct sorting in OHDSI Usagi. This is mainly used to help prioritizing
mapping to (manually) validated values vs not-validated (device) data.
}
\examples{
swapped <- swap_usagi_counts("numericitems_other")
swapped
}
