% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Small wrapper for translating using googleLanguageR.}
\usage{
translate(text, source = "nl", target = "en")
}
\arguments{
\item{text}{Character vector to translate}

\item{source}{Source language using two letter code.}

\item{target}{Target language using two letter code. Default: 'en' for English}
}
\value{
translated character vector
}
\description{
Small wrapper for translating using googleLanguageR.
}
\examples{
translation <- translate("Mooi weer is het vandaag")
}
