% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_concepts.R
\name{translate_concepts}
\alias{translate_concepts}
\title{Translate concepts}
\usage{
translate_concepts(concepts)
}
\arguments{
\item{concepts}{tibble containing concepts (column name: "source_concept")
and the number of occurrences in the database (column name: "count") of
those concepts.}
}
\value{
tibble containing concepts and their translation
}
\description{
Translates column_name into English using the Google Translate Cloud API for
matching using OHDSI Usagi.
}
\examples{
#' library(dplyr)
conn <- amstel::connect()
admissions <- tbl(conn, "admissions")
concepts <- admissions \%>\% group_by(source_concept=origin) \%>\% summarise(count=n()) \%>\% collect()
concepts <- translate_concepts(concepts)
}
