% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Creates \code{amstel} configuration file}
\usage{
create_config(is_missing = FALSE)
}
\arguments{
\item{is_missing}{Used internally to change the message based on if configuration file
is missing during initialization.
Default: FALSE.}
}
\description{
Creates the \code{config.yaml} configuration file in the default OS-defined
user configuration folder for setting up access to the databases and files
}
\examples{
amstel::create_config()
}
