% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_webapi_tables.R
\name{create_webapi_tables}
\alias{create_webapi_tables}
\title{Create OHDSI WebAPI tables}
\usage{
create_webapi_tables()
}
\description{
This function creates the required OHDSI WebAPI tables.
}
\details{
This function requires a configured \code{config.yaml} file created using
\code{amstel::create_config()}. It will use the connection configuration specified
in the \code{results} section to create WebAPI tables. WebAPI requires that
\code{amstel::etl()} (or the ETL background script) and \code{amstel::achilles()}
have already been run.
}
