% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_dashboard.R
\name{data_quality_dashboard}
\alias{data_quality_dashboard}
\title{Shows the OHDSI Data Quality Dashboard with the most recents results}
\usage{
data_quality_dashboard()
}
\description{
Shows the log created while executing
DataQualityDashboard::executeDqChecks()
}
