% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_dashboard.R
\name{execute_dqd_checks}
\alias{execute_dqd_checks}
\title{Evaluate Data Quality of AmsterdamUMCdb OMOP CDM after ETL.}
\usage{
execute_dqd_checks(custom_control = TRUE, check_names = c(), sql_only = FALSE)
}
\arguments{
\item{custom_control}{Specifies whether to apply the customized control files (thresholds)
for AmsterdamUMCdb
Default: TRUE}

\item{check_names}{Vector (\code{c()}) of data quality checks to run. Names can be found at:
https://github.com/OHDSI/DataQualityDashboard/blob/main/inst/csv/OMOP_CDMv5.4_Check_Descriptions.csv
Default: c()}
}
\description{
This function runs the tests of the OHDSI Data Quality
Dashboard.
Based on: https://ohdsi.github.io/DataQualityDashboard/articles/DataQualityDashboard.html
}
