% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_dashboard.R
\name{execute_dqd_checks}
\alias{execute_dqd_checks}
\title{Evaluate Data Quality of AmsterdamUMCdb OMOP CDM after ETL.}
\usage{
execute_dqd_checks()
}
\description{
This function runs the tests of the OHDSI Data Quality
Dashboard.
Based on: https://ohdsi.github.io/DataQualityDashboard/articles/DataQualityDashboard.html
}
